refactor: Extract StyledSVG
