+++ src/pages/ProjectText/ProjectVoiceOptions/SentenceInterval/index.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, MouseEvent, useCallback, useState } from 'react';
 
 import Label from '../../components/Label';
 import IconButton from '../../components/IconButton';
-import { StyledSVG } from '../../styles';
+import { StyledSVG } from '../../components/StyledSVG';
 import * as S from './styles';
 
 import { useTranslate } from '../../translate/hooks';
